Improve error showing by design in AddressChecker
